Fix statements
Add future_payments and and taxations objects properties
Note that the properties are incomplete. Chargify documentation do not
specify the schema of those objects. I am not able to infer the
properties from API call (my instance do not leverage those features). I
have a ticket opened with Chargify to get the documentation on those
object and will submitt another PR when I'll have the informations.