Singer io schema properties types (#30)
* fixes schema ids type

Internal Chargify ids are integers and not numbers

* fix schema cents types

In Chargify properties in cents are integers and not numbers

* fix schema unit price type

As per Chargify documentation the unit_price property is a string

* fix schema conversion_limit type

The conversion_limit is the number of time a coupon may b added to a
subscription. As for, it should be an integer (not clear from Chargify
documentation).

* fix schema invoice number type

As per the Chargify documentation, the number property is an string and
not a number.

* fix schema gateway_order_id

This is not very clear from Chargify documentation but I think that
this property should be a string rather than a number. The related
gateway_transaction_id is a string. We can not assume that the gateway
will send back an integer so string is a safer option.

* fix schema product_version_number property

Version numbers are integer in Chargify

* fix schema signup_revenue property

As per Chargify documentation, signup_revenue is a string

The revenue, formatted as a string of decimal separated
dollars and, cents, from the subscription signup
($50.00 would be formatted as 50.00)

* fix schema snap_day property

This property is an integer rather than a number.
This is the day of the month that the subcription will be
charged.

* fix schema net_terms property

This property is a number of day, hence an integer

* fix schema reference property

As per the Chargify documentation this property is a string.

* fix schema interval types

As per Chargify documentation, interval properties are integer

* fix schema accounting code property

As per Chargify documentation, accounting_code is a string

Co-authored-by: Damien Cirotteau <d@cirotteau.info>
Co-authored-by: Zach Harris <zharris@talend.com>